% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{flickr_afinn}
\alias{flickr_afinn}
\title{flickr_afinn
Added a column for the afinn sentiment for each photo}
\usage{
flickr_afinn(flickr_data = NULL)
}
\arguments{
\item{flickr_data}{}
}
\description{
flickr_afinn
Added a column for the afinn sentiment for each photo
}
