% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{flickr_nrc}
\alias{flickr_nrc}
\title{flikckr_nrc
Added a column for the count each of nrc sentiments for each photo}
\usage{
flickr_nrc(flickr_data = NULL)
}
\arguments{
\item{flickr_data}{}
}
\description{
flikckr_nrc
Added a column for the count each of nrc sentiments for each photo
}
