% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{paste2}
\alias{paste2}
\title{paste2
function to paste without copying NAs}
\usage{
paste2(..., sep = ", ")
}
\arguments{
\item{sep}{}
}
\description{
paste2
function to paste without copying NAs
}
