% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{paste2}
\alias{paste2}
\title{paste2
function to paste without copying NAs}
\usage{
paste2(..., sep = ", ")
}
\arguments{
\item{...}{what you wish to paste}

\item{sep}{how you want them seperated}
}
\value{
a pasted versions of what you requested
}
\description{
paste2
function to paste without copying NAs
}
\examples{
\dontrun{
flickr_raw$text <- paste2(flickr_raw$title, flickr_raw$description, flickr_raw$tags)
}

}
